fix: change numbers from int64 to int

+++ github/secret_scanning.go
@@ -17,7 +17,7 @@ import (
 type SecretScanningService service
 
 type SecretScanningAlert struct {
-	Number       *int64     `json:"number,omitempty"`
+	Number       *int       `json:"number,omitempty"`
 	CreatedAt    *Timestamp `json:"created_at,omitempty"`
 	URL          *string    `json:"url,omitempty"`
 	HTMLURL      *string    `json:"html_url,omitempty"`
@@ -37,10 +37,10 @@ type SecretScanningAlertLocation struct {
 
 type SecretScanningAlertLocationDetails struct {
 	Path        *string `json:"path,omitempty"`
-	Startline   *int64  `json:"start_line,omitempty"`
-	EndLine     *int64  `json:"end_line,omitempty"`
-	StartColumn *int64  `json:"start_column,omitempty"`
-	EndColumn   *int64  `json:"end_column,omitempty"`
+	Startline   *int    `json:"start_line,omitempty"`
+	EndLine     *int    `json:"end_line,omitempty"`
+	StartColumn *int    `json:"start_column,omitempty"`
+	EndColumn   *int    `json:"end_column,omitempty"`
 	BlobSHA     *string `json:"blob_sha,omitempty"`
 	BlobURL     *string `json:"blob_url,omitempty"`
 	CommitSHA   *string `json:"commit_sha,omitempty"`
